requirements.txt
Django==3.2.9
psycopg2-binary==2.9.2
Pillow==8.4.0
djangorestframework==3.12.4
python-dotenv==0.19.2
djoser==2.1.0